planner: address full join review comments
