+++ pkg/planner/core/casetest/fulljoin/BUILD.bazel
@@ -8,11 +8,12 @@ go_test(
         "main_test.go",
     ],
     flaky = True,
-    shard_count = 10,
+    shard_count = 13,
     deps = [
         "//pkg/config",
         "//pkg/domain",
         "//pkg/infoschema",
+        "//pkg/kv",
         "//pkg/meta/model",
         "//pkg/parser",
         "//pkg/parser/mysql",
@@ -23,10 +24,13 @@ go_test(
         "//pkg/planner/core/resolve",
         "//pkg/planner/core/rule",
         "//pkg/sessionctx",
+        "//pkg/store/mockstore",
         "//pkg/testkit",
+        "//pkg/testkit/external",
         "//pkg/testkit/testsetup",
         "//pkg/util/dbterror/plannererrors",
         "//pkg/util/mock",
+        "@com_github_pingcap_tipb//go-tipb",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
