+++ pkg/planner/core/physical_plan_test.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"fmt"
 	"math"
-	"strings"
 	"testing"
 
 	"github.com/pingcap/errors"
@@ -322,148 +321,6 @@ func TestJoinHintCompatibilityWithVariable(t *testing.T) {
 	tk.MustExec("select /*+ leading(t2), hash_join(t2) */ * from t t1 join t t2 join t t3 where t1.a = t2.a and t2.b = t3.b;")
 	res := tk.MustQuery("show warnings").Rows()
 	require.Equal(t, len(res) > 0, true)
-}
-
-func TestMPPFullOuterJoinToPB(t *testing.T) {
-	store := testkit.CreateMockStore(t, mockstore.WithMockTiFlash(2))
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_cost_model_version=2")
-	tk.MustExec("set @@tidb_allow_mpp=1")
-	tk.MustExec("set @@tidb_enforce_mpp=1")
-	tk.MustExec("set @@tidb_enable_full_outer_join=1")
-	tk.MustExec("create table t1 (a int, b int)")
-	tk.MustExec("create table t2 (a int, b int)")
-	tk.MustExec("alter table t1 set tiflash replica 1")
-	tk.MustExec("alter table t2 set tiflash replica 1")
-	for _, tableName := range []string{"t1", "t2"} {
-		tb := external.GetTableByName(t, tk, "test", tableName)
-		err := domain.GetDomain(tk.Session()).DDLExecutor().UpdateTableReplicaInfo(tk.Session(), tb.Meta().ID, true)
-		require.NoError(t, err)
-	}
-
-	sql := "select /*+ shuffle_join(t1, t2), read_from_storage(tiflash[t1, t2]) */ * from t1 full outer join t2 on t1.a = t2.a and t1.b > 1 and t2.b > 1"
-	rows := tk.MustQuery("explain format = 'brief' " + sql).Rows()
-	hasMPPFullOuterHashJoin := false
-	for _, row := range rows {
-		line := fmt.Sprint(row)
-		if strings.Contains(line, "HashJoin") && strings.Contains(line, "mpp[tiflash]") && strings.Contains(line, "full outer join") {
-			hasMPPFullOuterHashJoin = true
-			break
-		}
-	}
-	require.True(t, hasMPPFullOuterHashJoin)
-
-	stmt, err := parser.New().ParseOneStmt(sql, "", "")
-	require.NoError(t, err)
-	nodeW := resolve.NewNodeW(stmt)
-	plan, _, err := planner.Optimize(context.Background(), tk.Session(), nodeW, domain.GetDomain(tk.Session()).InfoSchema())
-	require.NoError(t, err)
-
-	var hashJoin *core.PhysicalHashJoin
-	flat := core.FlattenPhysicalPlan(plan, true)
-	for _, op := range flat.Main {
-		hj, ok := op.Origin.(*core.PhysicalHashJoin)
-		if !ok {
-			continue
-		}
-		hashJoin = hj
-		break
-	}
-	require.NotNil(t, hashJoin)
-	require.Len(t, hashJoin.LeftConditions, 1)
-	require.Len(t, hashJoin.RightConditions, 1)
-
-	pb, err := hashJoin.ToPB(tk.Session().GetBuildPBCtx(), kv.TiFlash)
-	require.NoError(t, err)
-	require.Equal(t, tipb.JoinType_TypeFullOuterJoin, pb.Join.JoinType)
-}
-
-func TestMPPNullEQJoinToPB(t *testing.T) {
-	store := testkit.CreateMockStore(t, mockstore.WithMockTiFlash(2))
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_cost_model_version=2")
-	tk.MustExec("set @@tidb_allow_mpp=1")
-	tk.MustExec("set @@tidb_enforce_mpp=1")
-	tk.MustExec("create table t1 (a int, b int)")
-	tk.MustExec("create table t2 (a int, b int)")
-	tk.MustExec("alter table t1 set tiflash replica 1")
-	tk.MustExec("alter table t2 set tiflash replica 1")
-	for _, tableName := range []string{"t1", "t2"} {
-		tb := external.GetTableByName(t, tk, "test", tableName)
-		err := domain.GetDomain(tk.Session()).DDLExecutor().UpdateTableReplicaInfo(tk.Session(), tb.Meta().ID, true)
-		require.NoError(t, err)
-	}
-
-	sql := "select /*+ shuffle_join(t1, t2), read_from_storage(tiflash[t1, t2]) */ * from t1 join t2 on t1.a <=> t2.a"
-	rows := tk.MustQuery("explain format = 'brief' " + sql).Rows()
-	hasMPPNullEQHashJoin := false
-	for _, row := range rows {
-		line := fmt.Sprint(row)
-		if strings.Contains(line, "HashJoin") && strings.Contains(line, "mpp[tiflash]") && strings.Contains(line, "nulleq(") {
-			hasMPPNullEQHashJoin = true
-			break
-		}
-	}
-	require.True(t, hasMPPNullEQHashJoin)
-
-	stmt, err := parser.New().ParseOneStmt(sql, "", "")
-	require.NoError(t, err)
-	nodeW := resolve.NewNodeW(stmt)
-	plan, _, err := planner.Optimize(context.Background(), tk.Session(), nodeW, domain.GetDomain(tk.Session()).InfoSchema())
-	require.NoError(t, err)
-
-	var hashJoin *core.PhysicalHashJoin
-	flat := core.FlattenPhysicalPlan(plan, true)
-	for _, op := range flat.Main {
-		hj, ok := op.Origin.(*core.PhysicalHashJoin)
-		if !ok {
-			continue
-		}
-		hashJoin = hj
-		break
-	}
-	require.NotNil(t, hashJoin)
-	require.Equal(t, []bool{true}, hashJoin.IsNullEQ)
-
-	pb, err := hashJoin.ToPB(tk.Session().GetBuildPBCtx(), kv.TiFlash)
-	require.NoError(t, err)
-	require.Equal(t, []bool{true}, pb.Join.GetIsNullEq())
-}
-
-func TestMPPFullOuterJoinWithoutShuffleHint(t *testing.T) {
-	store := testkit.CreateMockStore(t, mockstore.WithMockTiFlash(2))
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("set tidb_cost_model_version=2")
-	tk.MustExec("set @@tidb_allow_mpp=1")
-	tk.MustExec("set @@tidb_enforce_mpp=1")
-	tk.MustExec("set @@tidb_enable_full_outer_join=1")
-	tk.MustExec("set @@tidb_broadcast_join_threshold_count=1000000000")
-	tk.MustExec("set @@tidb_broadcast_join_threshold_size=1000000000")
-	tk.MustExec("create table t1 (a int, b int)")
-	tk.MustExec("create table t2 (a int, b int)")
-	tk.MustExec("alter table t1 set tiflash replica 1")
-	tk.MustExec("alter table t2 set tiflash replica 1")
-	for _, tableName := range []string{"t1", "t2"} {
-		tb := external.GetTableByName(t, tk, "test", tableName)
-		err := domain.GetDomain(tk.Session()).DDLExecutor().UpdateTableReplicaInfo(tk.Session(), tb.Meta().ID, true)
-		require.NoError(t, err)
-	}
-
-	sql := "select /*+ read_from_storage(tiflash[t1, t2]) */ * from t1 full outer join t2 on t1.a = t2.a and t1.b > 1 and t2.b > 1"
-	rows := tk.MustQuery("explain format = 'brief' " + sql).Rows()
-	hasMPPFullOuterHashJoin := false
-	for _, row := range rows {
-		line := fmt.Sprint(row)
-		if strings.Contains(line, "HashJoin") && strings.Contains(line, "mpp[tiflash]") && strings.Contains(line, "full outer join") {
-			hasMPPFullOuterHashJoin = true
-			break
-		}
-	}
-	require.True(t, hasMPPFullOuterHashJoin)
-	tk.MustQuery("show warnings").Check(testkit.Rows())
 }
 
 func TestHintAlias(t *testing.T) {
